Use bigint for invoice.invoice_number
Source: https://github.com/msupply-foundation/remote-server/commit/39fe6012c185767c7f41ce150ade3c0ee93890e9

+++ server/service/src/number.rs
@@ -8,7 +8,7 @@ pub fn next_number(
     connection: &StorageConnection,
     r#type: &NumberRowType,
     store_id: &str,
-) -> Result<i32, RepositoryError> {
+) -> Result<i64, RepositoryError> {
     // Should be done in transaction
     let next_number = connection.transaction_sync(|connection_tx| {
         let repo = NumberRowRepository::new(&connection_tx);
